refactor(colors): single source of truth shared with the browser
Colors were duplicated in TypeScript. Move the Ultralytics palette and Color
type from the native-only visualizer into a wasm-safe crate::colors module
(visualizer::color now re-exports it), add Color::to_hex(), and have the wasm
bindings attach the palette color to every detection (boxes, OBB, keypoints,
classify). The browser annotate() now uses engine-supplied colors and holds no
palette of its own, so person (class 0) is blue everywhere, matching the native
and Python output.

Also surface per-stage progress + full errors in the example page.
